Comment out urusan bersama otsuses menu item in layouts/menu.blade.php
The code diff shows that the nav-item for urusan bersama otsuses has been commented out in the layouts/menu.blade.php file. The code block for the menu item is still present, but it is now surrounded by HTML comment tags. This will prevent the menu item from being rendered in the application's navigation menu.