Update input value for 'tahun' field in monitoring\_penggunaans/fields.blade.php and monitoring\_penyalurans/fields.blade.php
This commit updates the value of the 'tahun' input field in the 'monitoring\_penggunaans/fields.blade.php' and 'monitoring\_penyalurans/fields.blade.php' views to use the 'tahun' property of the '$monitoringPenggunaan' and '$monitoringPenyaluran' objects, respectively. Previously, both views used the '$tahun' variable, which is not defined in the context of these views.