Change SuratTugas query to include 'jenis\_penugasan' filter as 'Evaluasi'
The commit changes the SuratTugas query in the 'index' method of the 'EvaluasiLaporanController' class to include a 'where' clause that filters the results based on 'jenis\_penugasan' equal to 'Evaluasi'. This is in addition to the existing 'where' clause that filters the results based on 'jenis\_tkd' equal to the current session's 'jenis\_tkd'.